Split data values sets and analytics fetch data
