+++ lib/orbf/rules_engine.rb
@@ -46,6 +46,8 @@ require_relative "./rules_engine/fetch_data/datasets/compute_orgunits"
 require_relative "./rules_engine/fetch_data/datasets/compute_data_elements"
 require_relative "./rules_engine/fetch_data/datasets/compute_datasets"
 require_relative "./rules_engine/fetch_data/resolve_arguments"
+require_relative "./rules_engine/fetch_data/fetch_data_value_sets"
+require_relative "./rules_engine/fetch_data/fetch_data_analytics"
 require_relative "./rules_engine/fetch_data/fetch_data"
 require_relative "./rules_engine/fetch_data/orgunit_facts"
 require_relative "./rules_engine/fetch_data/pyramid_factory"
